+++ src/meta/validator.spec.ts
@@ -344,6 +344,7 @@ function mockEntry(tagName: string, stub = {}): MetaElement {
 		deprecated: false,
 		void: false,
 		transparent: false,
+		deprecatedAttributes: [],
 		permittedContent: [],
 		permittedDescendants: [],
 		permittedOrder: [],